+++ packages/terra-clinical-result/src/terra-dev-site/doc/example/clinical-result-blood-pressure/ClinicalResultBloodPressureChangeSize.jsx
