+++ packages/terra-clinical-result/src/terra-dev-site/doc/example/flowsheet-result-cell/FlowsheetResultCellStatusEnteredInError.jsx
