+++ packages/terra-clinical-result/src/terra-dev-site/doc/example/result-name-header-cell/ResultNameHeaderCell.jsx
