+++ packages/terra-clinical-result/src/terra-dev-site/doc/example/result-time-header-cell/ResultTimeHeaderCell.jsx
